pyproject.toml
[project]
name = "sales_data"
version = "0.0.1"
authors = [{ name = "huthaifa.jadallah@hotmail.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    # Any dependencies for jobs and pipelines in this project can be added here
    # See also https://docs.databricks.com/dev-tools/bundles/library-dependencies
    #
    # LIMITATION: for pipelines, dependencies are cached during development;
    # add dependencies to the 'environment' section of your pipeline.yml file instead
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff",
    "databricks-dlt",
    "databricks-connect>=15.4,<15.5",
    "pyspark>=3.4.0",
    "ipykernel",
]

[project.scripts]
main = "sales_data.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
